added condition if there is a profile_path on database before checking if the file exists

+++ app/Http/Controllers/UsersController.php
@@ -18,7 +18,7 @@ class UsersController extends Controller
     public function remove(Account $account)
     {
         // Delete Profile Pic
-        if (file_exists(public_path('images/accounts/' . $account->profile_path))){
+        if (isset($account->profile_path) && file_exists(public_path('images/accounts/' . $account->profile_path))){
             unlink(public_path('images/accounts/' . $account->profile_path));
         }
